ReSharper refactor
