+++ TagsCloudVisualization/TagsCloudVisualization/Helpers/RectangleCreator.cs
@@ -1,14 +1,13 @@
 ﻿using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Helpers;
+
+public static class RectangleCreator
 {
-	public static class RectangleCreator
+	public static Rectangle GetRectangle(Point rectangleCenter, Size rectangleSize)
 	{
-		public static Rectangle GetRectangle(Point rectangleCenter, Size rectangleSize)
-		{
-			var x = rectangleCenter.X - rectangleSize.Width / 2;
-			var y = rectangleCenter.Y - rectangleSize.Height / 2;
-			return new Rectangle(new Point(x, y), rectangleSize);
-		}
+		var x = rectangleCenter.X - rectangleSize.Width / 2;
+		var y = rectangleCenter.Y - rectangleSize.Height / 2;
+		return new Rectangle(new Point(x, y), rectangleSize);
 	}
 }
