+++ TagsCloudVisualization/TagsCloudVisualization/ImageSaver.cs
@@ -1,23 +1,25 @@
 ﻿using System.Drawing;
 using System.Drawing.Imaging;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization;
+
+public class ImageSaver
 {
-	public class ImageSaver
+	public ImageSaver(string directory, ImageFormat format)
 	{
-		public ImageFormat Format { get; }
-		public string Directory { get; }
+		Format = format;
+		Directory = directory;
 
-		public ImageSaver(string directory, ImageFormat format)
-		{
-			Format = format;
-			Directory = directory;
-		}
+		System.IO.Directory.CreateDirectory(directory);
+	}
 
-		public void Save(Image image, string name)
-		{
-			var path = Directory + name + "." + Format.ToString().ToLower();
-			image.Save(path, Format);
-		}
+	public ImageFormat Format { get; }
+	public string Directory { get; }
+
+	public string Save(Image image, string name)
+	{
+		var path = $"{Directory}\\{name}.{Format.ToString().ToLower()}";
+		image.Save(path, Format);
+		return path;
 	}
 }
