+++ TagsCloudVisualization/TagsCloudVisualization/Spirals/ArchimedeanSpiral.cs
@@ -1,36 +1,41 @@
 ﻿using System.Drawing;
+using TagsCloudVisualization.Helpers;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization.Spirals;
+
+public class ArchimedeanSpiral : ISpiral
 {
-	public class ArchimedeanSpiral : ISpiral
+	private readonly Point center;
+
+	public ArchimedeanSpiral(Point center, double spiralStepInPixels, double angleStepInDegrees)
 	{
-		public double SpiralStepInPixels { get; }
+		if (spiralStepInPixels == 0)
+			throw new ArgumentException("Spiral step should not be equals zero");
+		if (angleStepInDegrees == 0)
+			throw new ArgumentException("Angle step should not be equals zero");
 
-		public double AngleStepInRadians { get; }
+		this.center = center;
+		SpiralStepInPixels = spiralStepInPixels;
+		AngleStepInRadians = angleStepInDegrees * Math.PI / 180;
+		CurrentAngleInRadians = 0;
+	}
 
-		public double CurrentAngleInRadians { get; private set; }
+	public double SpiralStepInPixels { get; }
 
-		public ArchimedeanSpiral(Point center, double spiralStepInPixels, double angleStepInDegrees)
-		{
-			this.center = center;
-			SpiralStepInPixels = spiralStepInPixels;
-			AngleStepInRadians = angleStepInDegrees * Math.PI / 180;
-			CurrentAngleInRadians = 0;
-		}
+	public double AngleStepInRadians { get; }
 
-		public Point GetNextPoint()
-		{
-			var polarRadius = SpiralStepInPixels * CurrentAngleInRadians;
-			var x = polarRadius * Math.Cos(CurrentAngleInRadians);
-			var y = polarRadius * Math.Sin(CurrentAngleInRadians);
-			
-			var localPosition = new Point(Convert.ToInt32(x), Convert.ToInt32(y));
+	public double CurrentAngleInRadians { get; private set; }
 
-			CurrentAngleInRadians += AngleStepInRadians;
+	public Point GetNextPoint()
+	{
+		var polarRadius = SpiralStepInPixels / (Math.PI * 2) * CurrentAngleInRadians;
+		var x = polarRadius * Math.Cos(CurrentAngleInRadians);
+		var y = polarRadius * Math.Sin(CurrentAngleInRadians);
 
-			return localPosition.Plus(center);
-		}
+		var localPosition = new Point(Convert.ToInt32(x), Convert.ToInt32(y));
 
-		private readonly Point center;
+		CurrentAngleInRadians += AngleStepInRadians;
+
+		return localPosition.Plus(center);
 	}
 }
